refactor: deconstruct discovery/providers.py module
- Decomposed the monolithic `discovery/providers.py` file into a modular package `discovery/providers/`.
- Created individual files for each provider type: `appliance.py`, `camera.py`, `device.py`, `mt_sensor.py`, `switch.py`, `uplink.py`, `wireless.py`.
- Created `__init__.py` to export all providers, maintaining backward compatibility.
- Reduced individual file sizes to well under 300 lines, satisfying Agent Readiness standards.
- Verified with `pytest` and `mypy`.
