+++ custom_components/meraki_ha/discovery/providers/wireless.py
@@ -1,0 +1,114 @@
+"""Wireless Radio Provider."""
+from __future__ import annotations
+
+import logging
+from typing import TYPE_CHECKING, Any
+
+from homeassistant.components.sensor import SensorEntityDescription
+from homeassistant.const import EntityCategory
+
+from ...sensor.device.wireless_radio import MerakiWirelessRadioSensor
+
+if TYPE_CHECKING:
+    from homeassistant.config_entries import ConfigEntry
+    from homeassistant.helpers.entity import Entity
+
+    from ...coordinator import MerakiDataUpdateCoordinator
+    from ...core.models.device import MerakiDevice
+
+_LOGGER = logging.getLogger(__name__)
+
+
+class WirelessRadioProvider:
+    """Provider for wireless radio entities."""
+
+    @staticmethod
+    def get_entities(
+        coordinator: MerakiDataUpdateCoordinator,
+        device: MerakiDevice,
+        config_entry: ConfigEntry,
+        **kwargs: Any,
+    ) -> list[Entity]:
+        """Get entities."""
+        if not device.serial or not device.wireless_radio_settings:
+            return []
+
+        entities: list[Entity] = []
+        settings = device.wireless_radio_settings
+
+        # 2.4GHz Channel
+        if "twoFourGhzSettings" in settings:
+            entities.append(
+                MerakiWirelessRadioSensor(
+                    coordinator,
+                    device,
+                    config_entry,
+                    SensorEntityDescription(
+                        key="2.4ghz_channel",
+                        name="2.4GHz channel",
+                        icon="mdi:wifi",
+                        entity_category=EntityCategory.DIAGNOSTIC,
+                    ),
+                    "twoFourGhzSettings",
+                    "channel",
+                )
+            )
+
+        # 5GHz Channel
+        if "fiveGhzSettings" in settings:
+            entities.append(
+                MerakiWirelessRadioSensor(
+                    coordinator,
+                    device,
+                    config_entry,
+                    SensorEntityDescription(
+                        key="5ghz_channel",
+                        name="5GHz channel",
+                        icon="mdi:wifi",
+                        entity_category=EntityCategory.DIAGNOSTIC,
+                    ),
+                    "fiveGhzSettings",
+                    "channel",
+                )
+            )
+
+        # 2.4GHz Target Power
+        if "twoFourGhzSettings" in settings:
+            entities.append(
+                MerakiWirelessRadioSensor(
+                    coordinator,
+                    device,
+                    config_entry,
+                    SensorEntityDescription(
+                        key="2.4ghz_target_power",
+                        name="2.4GHz target power",
+                        native_unit_of_measurement="dBm",
+                        icon="mdi:transmission-tower",
+                        entity_category=EntityCategory.DIAGNOSTIC,
+                    ),
+                    "twoFourGhzSettings",
+                    "targetPower",
+                )
+            )
+
+        # 5GHz Target Power (Generic "Target power" sensor for backward
+        # compatibility/requested name)
+        if "fiveGhzSettings" in settings:
+            entities.append(
+                MerakiWirelessRadioSensor(
+                    coordinator,
+                    device,
+                    config_entry,
+                    SensorEntityDescription(
+                        key="target_power",
+                        name="Target power",
+                        native_unit_of_measurement="dBm",
+                        icon="mdi:transmission-tower",
+                        entity_category=EntityCategory.DIAGNOSTIC,
+                    ),
+                    "fiveGhzSettings",
+                    "targetPower",
+                )
+            )
+
+        return entities
